backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Yohann MARTIN <mimichackeur@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^2.0.2"
Flask-Cors = "^3.0.10"
Flask-JWT-Extended = "^4.3.1"
flask-marshmallow = "^0.14.0"
Flask-Migrate = "^3.1.0"
Flask-RESTful = "^0.3.9"
PyJWT = "^2.3.0"
psycopg2 = "^2.9.3"
marshmallow-sqlalchemy = "^0.27.0"
passlib = "^1.7.4"
marshmallow-enum = "^1.5.1"
Flask-Mail = "^0.9.1"
python-dotenv = "^0.19.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
